8th HW: second edition

+++ src/main/java/com/playtika/automation/carshop/domain/CarSaleDetails.java
@@ -3,13 +3,11 @@ package com.playtika.automation.carshop.domain;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
-import javax.validation.constraints.NotNull;
-
 @Data
 @AllArgsConstructor
 public class CarSaleDetails {
     private long id;
     private Car car;
-    private double price;
+    private int price;
     private String contacts;
 }
