+++ src/main/java/com/playtika/automation/carshop/service/CarService.java
@@ -10,5 +10,5 @@ public interface CarService {
     Collection<CarSaleDetails> getAllCars();
     Map<String,Object> getCarDetailsById(long id);
     void deleteCarById(long id);
-    CarSaleDetails addCar(Car car, double price, String contacts);
+    CarSaleDetails addCar(Car car, int price, String contacts);
 }
